rename prismarine folder
