+++ www/src/player.ts
@@ -4,7 +4,7 @@ import {PickAt} from "./pickat.js";
 import {Instrument_Hand} from "./instrument/hand.js";
 import {BLOCK} from "./blocks.js";
 import {PLAYER_DIAMETER, DEFAULT_SOUND_MAX_DIST, PLAYER_STATUS } from "./constant.js";
-import {PrismarinePlayerControl, PHYSICS_TIMESTEP} from "./vendors/prismarine-physics/using.js";
+import {PrismarinePlayerControl, PHYSICS_TIMESTEP} from "./prismarine-physics/using.js";
 import {PlayerControl, SpectatorPlayerControl} from "./spectator-physics.js";
 import {PlayerInventory} from "./player_inventory.js";
 import { PlayerWindowManager } from "./player_window_manager.js";
