+++ www/src/vendors/prismarine-physics/features.ts
@@ -1,5 +1,5 @@
 // Load supported block styles
 export let features = [];
-await fetch('/src/vendors/prismarine-physics/lib/features.json').then(response => response.json()).then(json => {
+await fetch('/src/prismarine-physics/lib/features.json').then(response => response.json()).then(json => {
     features = json;
 });
